refactor: Move `ErrorScreen` to `app/View/Error/ErrorScreen``
